requirements.txt
flask
flask-socketio
python-dotenv
psycopg2-binary
gevent
gevent-websocket